pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.28",
    "openai>=1.98.0",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
    "tabulate>=0.9.0",
]
